Pull providers into single file
