--- corehq/ex-submodules/casexml/apps/phone/data_providers/case/__init__.py
@@ -1,1 +1,0 @@
-from .providers import *
